requirements.txt
langchain==0.1.14
langchain-community==0.0.38
langchain-openai==0.1.8
langchain-core==0.1.48
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
unstructured==0.12.5
pdf2image==1.16.3
pillow==10.1.0
pytesseract==0.3.10
pypdf==3.17.1
python-docx==0.8.11
openpyxl==3.11.0

# Embeddings and Vector DB
openai==1.3.9
sentence-transformers==2.2.2
chromadb==0.4.24
faiss-cpu==1.7.4
qdrant-client==1.7.0
langchain-qdrant==0.1.0

# Database and Storage
pymongo==4.6.0
sqlalchemy==2.0.23

# Utilities
numpy==1.24.3
pandas==2.1.4
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
